TWEAK: Disable FromObject optimizations
In some usages it caused ObjectDisposedException.

if (operand_1.GetObjectType() == PdfObjectType.Integer) {
    using var operand_1_integer = PdfIntegerObject.FromObject(operand_1);
    text_translate_y = operand_1_integer.IntegerValue;
}

if (operand_1.GetObjectType() == PdfObjectType.Real) {
    using var operand_1_real = PdfRealObject.FromObject(operand_1);
    text_translate_y = operand_1_real.Value;
}

The second call of GetObjectType() would crash,
since is was disposed within the first block.

It is much cleaner to duplicate the object at the level of .NET
as both of them would be referencing the same native object.