TWEAK: Update linux native paths
Since .NET 8, we should no longer use distro-specific builds.
In either case we embed all the dependencies within the native package,
so we should not care what linux distro this is.

Currently we also embed C/C++ runtime,
which increases the binary dramatically.
This might be changed in the future,
however for the time being we are good.